Create couplingAndDecoupling.xml
TruckState=>have various states for an individual truck wanting to join the platoon. It interacts with other components of the system through channels like move, join, finish, continues_to_move_platoon, continues_to_move_indi, joined, not_joined, maintain_distance, leave, and restart_platoon.

Watchdog=> Works as a monitoring component for coupling. Time is set to 10, if the truck doesnot join the in the given time, platoon resets.

Platoon_limit=> it limits the number of trucks in platoon to 4 and keep track of trucks joining and leaving.

Platoon=> Accepts the request for joining  and handles the channels like  join, leave, and restart_platoon.